light: Allocate permanent stack space for the kernel
Move the per-processor kernel stack to memory location 0xFFFFFFFF_F8000000 and
make it 4 KiB in size. Note that we may change these settings in the future.

In order to achieve the above, due to limited stack size in the light
environment, we use an intermediate stack inside framebuffer memory for the BSP.